requirements.txt
flask
tensorflow
pillow
numpy
mysql-connector-python
werkzeug
gunicorn
